[Step4] refactor: 상태에서 카드를 꺼내는 함수의 반환형을 Nullable 하도록 변경
