+++ src/main/kotlin/blackjack/domain/state/Started.kt
@@ -1,5 +1,6 @@
 package blackjack.domain.state
 
+import blackjack.domain.Card
 import blackjack.domain.Hand
 
 abstract class Started(
@@ -8,5 +9,17 @@ abstract class Started(
 
     override fun getSum(): Int {
         return hand.getSum()
+    }
+
+    override fun receive(card: Card) {
+        hand.receive(card)
+    }
+
+    override fun init(cards: List<Card>) {
+        hand.init(cards)
+    }
+
+    override fun cards(): Hand {
+        return hand
     }
 }
